Extract errors and readSafe() from collector
Allow for these to be reused by other modules.
