+++ lib/collector.js
@@ -2,73 +2,16 @@
 
 const path = require('path')
 
-const ExtendableError = require('es6-error')
-const gfs = require('graceful-fs')
 const parseJson5 = require('json5').parse
 const cloneDeep = require('lodash.clonedeep')
 
-function readSafe (source, cache) {
-  if (cache && cache.files && cache.files.has(source)) {
-    return cache.files.get(source)
-  }
-
-  const promise = new Promise((resolve, reject) => {
-    gfs.readFile(source, (err, contents) => {
-      if (err) {
-        if (err.code === 'ENOENT') {
-          resolve(null)
-        } else {
-          reject(err)
-        }
-      } else {
-        resolve(contents)
-      }
-    })
-  })
-
-  if (cache && cache.files) {
-    cache.files.set(source, promise)
-  }
-  return promise
-}
-
-class SourceError extends ExtendableError {
-  constructor (message, source, parent) {
-    super(`${source}: ${message}`)
-    this.source = source
-    this.parent = parent || null
-  }
-}
-
-class NoSourceFileError extends SourceError {
-  constructor (source) {
-    super('No such file', source)
-  }
-}
-
-class ParseError extends SourceError {
-  constructor (source, parent) {
-    super(`Error while parsing — ${parent.message}`, source, parent)
-  }
-}
-
-class InvalidFileError extends SourceError {
-  constructor (source) {
-    super('Not a proper configuration file', source)
-  }
-}
-
-class ExtendsError extends SourceError {
-  constructor (source, clause, parent) {
-    super(`Couldn't resolve extends clause: ${clause}`, source, parent)
-    this.clause = clause
-  }
-}
+const errors = require('./errors')
+const readSafe = require('./readSafe')
 
 function makeValid (source, options) {
   // Arrays are never valid options.
   if (Array.isArray(options)) {
-    throw new InvalidFileError(source)
+    throw new errors.InvalidFileError(source)
   }
 
   // Force options to be an object. Babel itself ignores falsy values when
@@ -83,7 +26,7 @@ function parseFile (source, buffer) {
   try {
     options = parseJson5(buffer.toString('utf8'))
   } catch (err) {
-    throw new ParseError(source, err)
+    throw new errors.ParseError(source, err)
   }
 
   return makeValid(source, options)
@@ -95,7 +38,7 @@ function parsePackage (source, buffer) {
     const pkg = JSON.parse(buffer.toString('utf8'))
     options = pkg && pkg.babel
   } catch (err) {
-    throw new ParseError(source, err)
+    throw new errors.ParseError(source, err)
   }
 
   return makeValid(source, options)
@@ -173,7 +116,7 @@ function resolveFile (source, cache) {
     .then(contents => {
       // The file *must* exist. Causes a proper error to be propagated to
       // where "extends" directives are resolved.
-      if (!contents) throw new NoSourceFileError(source)
+      if (!contents) throw new errors.NoSourceFileError(source)
 
       return new Config(source, true, parseFile(source, contents))
     })
@@ -238,7 +181,7 @@ class Collector {
           .then(extendsPointer => (config.extendsPointer = extendsPointer))
           .catch(err => {
             if (err.name === 'NoSourceFileError') {
-              throw new ExtendsError(config.source, extendsClause, err)
+              throw new errors.ExtendsError(config.source, extendsClause, err)
             }
 
             throw err
